# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    get:
      operationId: ArtistController_getAllArtists
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtistResponse'
      tags:
        - Artist
    post:
      operationId: ArtistController_createArtist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '400':
          description: body does not contain required fields
      tags:
        - Artist
  /artist/{id}:
    get:
      operationId: ArtistController_getArtistById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '400':
          description: id must be valid uuid
        '404':
          description: record not found
      tags:
        - Artist
    put:
      operationId: ArtistController_updateArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '400':
          description: id must be valid uuid or body does not contain required fields
        '404':
          description: record not found
      tags:
        - Artist
    delete:
      operationId: ArtistController_deleteArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: artist is found and deleted
        '400':
          description: id must be valid uuid
        '404':
          description: record not found
      tags:
        - Artist
...
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addFavoriteArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
        '400':
          description: id must be valid uuid
        '422':
          description: record whith this id doesn't exist
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_deleteFavoriteArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: id must be valid uuid
        '404':
          description: record not found
      tags:
        - Favorites
...
info:
  title: REST-Service-API
  description: REST API Documentation
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
...
    ArtistResponse:
      type: object
      properties:
        id:
          type: string
          example: 2981b70c-d670-4f9c-adaa-236dbf0edbe6
          description: uuid v4
        name:
          type: string
          example: John Lennon
        grammy:
          type: boolean
          example: true
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: John Lennon
        grammy:
          type: boolean
          example: true
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          example: John Lennon
        grammy:
          type: boolean
          example: false
      required:
        - name
        - grammy
...